divide tests into test_item.py and test_phone.py
