+++ tests/test_item.py
@@ -1,7 +1,6 @@
 """Здесь надо написать тесты с использованием pytest для модуля item."""
 import pytest
 from src.item import Item
-from src.phone import Phone
 
 
 @pytest.fixture
@@ -49,25 +48,5 @@ def test_repr():
     assert repr(item1) == "Item('Телевизор', 45000, 4)"
 
 
-phone1 = Phone("iPhone 14", 120_000, 5, 2)
 
-
-def test_phone_init():
-    assert phone1.name == "iPhone 14"
-    assert phone1.price == 120_000
-    assert phone1.quantity == 5
-    assert phone1.number_of_sim == 2
-
-
-def test_phone_str():
-    assert str(phone1) == 'iPhone 14'
-
-
-def test_phone_repr():
-    assert repr(phone1) == "Phone('iPhone 14', 120000, 5, 2)"
-
-
-def test_add():
-    assert item1 + phone1 == 9
-    assert phone1 + phone1 == 10
